+++ src/ByteDev.Testing/TestApiKey.cs
@@ -4,34 +4,34 @@ using System.Reflection;
 namespace ByteDev.Testing
 {
     /// <summary>
-    /// Represents a test connection string setting.
+    /// Represents a test API key setting.
     /// </summary>
-    public class TestConnectionString : TestSingleSetting
+    public class TestApiKey : TestSingleSetting
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="T:ByteDev.Testing.TestConnectionString" /> class.
+        /// Initializes a new instance of the <see cref="T:ByteDev.Testing.TestApiKey" /> class.
         /// </summary>
-        public TestConnectionString()
+        public TestApiKey()
         {
         }
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="T:ByteDev.Testing.TestConnectionString" /> class.
+        /// Initializes a new instance of the <see cref="T:ByteDev.Testing.TestApiKey" /> class.
         /// Adds default file paths based on the containing assembly's name.
         /// </summary>
         /// <param name="containingAssembly">Containing test assembly that is consuming the setting.</param>
         /// <exception cref="T:System.ArgumentNullException"><paramref name="containingAssembly" /> is null.</exception>
-        public TestConnectionString(Assembly containingAssembly)
+        public TestApiKey(Assembly containingAssembly)
         {
             if (containingAssembly == null)
                 throw new ArgumentNullException(nameof(containingAssembly));
 
-            FilePaths = DefaultFilePaths.GetDefaultFilePaths(DefaultFileName.GetConnStringFileName(containingAssembly));
+            FilePaths = DefaultFilePaths.GetDefaultFilePaths(DefaultFileName.GetApiKeyFileName(containingAssembly));
         }
-        
+
         public override string GetValue()
         {
-            return GetSettingValue("Could not find test connection string.");
+            return GetSettingValue("Could not find test API key.");
         }
     }
 }